Feature/user handling

+++ osa4/bloglist/controllers/blogs.js
@@ -1,9 +1,17 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  try {
+    const blogs = await Blog
+      .find({}).populate('user', { username: 1, name: 1 })
+
+    response.json(blogs)
+  } catch (e) {
+    throw e
+  }
 })
 
 blogRouter.get('/:id', async (request, response) => {
@@ -15,38 +23,70 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-
 blogRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
-  })
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
 
-  const savedBlog = await blog.save()
-  response.status(201).json(savedBlog)
+    const blog = new Blog({
+      ...body,
+      user: user._id
+    })
+    const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.status(201).json(savedBlog)
+  } catch (e) {
+    throw e
+  }
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  try {
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({
+        error: 'token missing or invalid'
+      })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const isUserAuthorized = blog.user.toString() === user.id.toString()
+
+    if (!(request.token || isUserAuthorized)) {
+      return response.status(404).json({ error: 'not authorized' })
+    }
+
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } catch (e) {
+    throw e
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
   const body = request.body
 
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
-  }
+  try {
+    const blog = {
+      ...body
+    }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-  response.json(updatedBlog).end()
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+    response.json(updatedBlog).end()
+  } catch (e) {
+    throw e
+  }
 })
 
 module.exports = blogRouter
